Version 0.8
This is an update to the plugin, redesigning the code to replicate the
Phaser.Gamepad interface as close as possible.

This version introduces both Phaser.OuyaGamepad and Phaser.OuyaSinglePad
objects as copies of their Gamepad equivalents, but with slight changes
to make the CocoonJS system.

There is now a demo too, matching the Gamepad example from the Phaser
library.